[Feat] #16  마이페이지 - 내 쿠폰함 화면 쿠폰 사용 API 연동

+++ app/src/main/java/com/project/drinkly/api/response/coupon/CouponListResponse.kt
@@ -1,8 +1,8 @@
 package com.project.drinkly.api.response.coupon
 
 data class CouponListResponse(
-    val id: Int,
-    val memberId: Int,
+    val id: Long,
+    val memberId: Long,
     val type: String,
     val status: String,
     val used: Boolean,
